pipeline/requirements.txt
dagster==1.9.3
dagit==1.9.3
dagster-postgres==0.25.3
dagster-pyspark==0.25.3

polars==1.17.1
pyarrow==18.1.0
minio==7.2.12
SQLAlchemy==2.0.36
cryptography==44.0.0
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pytest==8.3.4
pyspark==3.5.4
geopandas==1.0.1
clickhouse-driver==0.2.9
dagster-dbt==0.25.3
dbt-clickhouse==1.8.7
python-dateutil==2.9.0
great-expectations==0.18.22
